TP_Microservices_Mini_Projet.md/mini-projet/health-service/requirements.txt
flask~=2.1.2
Werkzeug==2.2.3
Flask-Cors~=3.0.10
PyJWT~=2.4.0
requests~=2.31.0
gunicorn~=20.1.0
